+++ optimizedConvolution.c
@@ -1,9 +1,7 @@
 #include "dto.h"
 
-/* Not Optimized Yet */
-
 void convolveOptimized(FilesDTO input) {
-    int sum = 0;
+    float sum = 0;
     for (int i = input.deg*input.deg-1; i >= 0 ; i--) {
         sum += input.mat[i];
     }
